port_operations_app/backend/requirements.txt
Django==4.2.21
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.10
Pillow==11.2.1
django-cors-headers==4.7.0
python-decouple==3.8 